internals: re-arrange implementation
This moves the DirEntry and Error types out into their own separate
modules.

This is prep work to (hopefully) make the impending refactoring (or
more likely, rewrite) more palatable.
